Move files to MVC structure folders
